rename and extract groupNodesByNamespace
